Improve Node#properties=
Node#properties= takes a hash and sets the properties of an object. It used to create a child node if a hash was given as a property value. This worked, but fell short when it came to updating a node.

Now, instead of turning hash property values into nodes, it turns NodeBlueprint property values into nodes. This let's the node be fully represented -- path, primary type, and properties. If the node already exists, it is updated in place.