Add Node#move method
Also, Node#save now calls save on the session, not the node. This was necessary otherwise a ConstraintViolationException was thrown — changes were made on the node via the session when it is moved. Instead of doing something fancy and complicated to preserve calling save on the node most of the time, I decided to call it on session. This is now what the JCR spec recommends.